Added landing and login page
