+++ app/routes/logout.tsx
@@ -1,6 +1,6 @@
 import type { ActionFunction } from "@remix-run/node";
 import { redirect } from "@remix-run/node";
-import { logout } from "~/session.server";
+import { logout } from "~/server/session.server";
 
 export const action: ActionFunction = async ({ request }) => {
   return logout(request);
