+++ app/routes/notes/$noteId.tsx
@@ -1,9 +1,9 @@
 import type { ActionFunction, LoaderArgs } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { Form, useLoaderData } from "@remix-run/react";
-import type { Note } from "~/models/note.server";
-import { deleteNote, getNote } from "~/models/note.server";
-import { requireUserId } from "~/session.server";
+import type { Note } from "~/server/models/note.server";
+import { deleteNote, getNote } from "~/server/models/note.server";
+import { requireUserId } from "~/server/session.server";
 import invariant from "tiny-invariant";
 
 type LoaderData = {
