+++ app/routes/notes/new.tsx
@@ -1,8 +1,8 @@
 import type { ActionFunction } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { Form } from "@remix-run/react";
-import { createNote } from "~/models/note.server";
-import { requireUserId } from "~/session.server";
+import { createNote } from "~/server/models/note.server";
+import { requireUserId } from "~/server/session.server";
 
 export const action: ActionFunction = async ({ request }) => {
   const userId = await requireUserId(request);
