+++ app/server/supabase.config.ts
@@ -1,0 +1,18 @@
+import { createClient } from "@supabase/supabase-js";
+import invariant from "tiny-invariant";
+
+// Abstract this away
+const supabaseUrl = process.env.SUPABASE_URL;
+const supabaseAnonKey = process.env.SUPABASE_ANON_KEY;
+
+invariant(
+  supabaseUrl,
+  "SUPABASE_URL must be set in your environment variables."
+);
+
+invariant(
+  supabaseAnonKey,
+  "SUPABASE_ANON_KEY must be set in your environment variables."
+);
+
+export const supabase = createClient(supabaseUrl, supabaseAnonKey);
