+++ app/utils.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { useMatches } from "@remix-run/react";
-import type { User } from "./models/user.server";
+import type { User } from "./server/models/user.server";
 
 export function useMatchesData(id: string) {
   const matchingRoutes = useMatches();
