+++ package.json
@@ -31,7 +31,7 @@
     "@remix-run/netlify": "^1.6.8",
     "@remix-run/node": "^1.6.8",
     "@remix-run/react": "^1.6.8",
-    "@supabase/supabase-js": "^1.31.1",
+    "@supabase/supabase-js": "^2.1.1",
     "@testing-library/cypress": "^8.0.3",
     "bcryptjs": "^2.4.3",
     "cypress": "^10.4.0",
@@ -43,6 +43,7 @@
     "@remix-run/dev": "^1.6.8",
     "@remix-run/eslint-config": "^1.6.8",
     "@remix-run/serve": "^1.6.8",
+    "@tailwindcss/typography": "^0.5.8",
     "@types/bcryptjs": "^2.4.2",
     "@types/eslint": "^8.4.1",
     "@types/react": "^17.0.24",
